refactor LineNumberDisplay class
The widget is set to be non-interactive. The appearance set to be close to QLabel.
Accumulation of line numbers made throug QTextStream since it is slightly faster.
The right-side alignment has to applied every time the text is updated.
The width is calculated as width of the longest character of the chosen font
times number of necessary digits with extra two digits.