improve Note class
Tabs are named as "Tab 1", "Tab 2" and so on instead of "untitled".
The icon of the button to add new tab is changed.
New button to run the script is added when qtisas is built with Python.
The script can be run as a whole, running and evaluating parts of the script
is no longer supported.
The output is redirected to qtisas script console, when the run is triggered,
the scripting console gets activated automatically.
Keyboard shortcuts are also adjusted for consistency with script window.